Split the test into seperate class.
