+++ src/test/java/io/stargate/sgv2/jsonapi/api/v1/VectorizeSearchIntegrationTest.java
@@ -7,11 +7,9 @@ import static org.hamcrest.Matchers.*;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 import io.restassured.http.ContentType;
-import io.stargate.sgv2.jsonapi.config.constants.HttpConstants;
 import io.stargate.sgv2.jsonapi.exception.ErrorCode;
 import io.stargate.sgv2.jsonapi.testresource.DseTestResource;
 import java.util.List;
-import java.util.Map;
 import org.junit.jupiter.api.ClassOrderer;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Nested;
@@ -19,7 +17,6 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestClassOrder;
 import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable;
 
 @QuarkusIntegrationTest
 @QuarkusTestResource(DseTestResource.class)
@@ -1229,142 +1226,6 @@ public class VectorizeSearchIntegrationTest extends AbstractNamespaceIntegration
           .body("data.documents[0].$vector", is(notNullValue()))
           .body("data.documents[0].$vector", contains(0.1f, 0.15f, 0.3f, 0.12f, 0.05f, 0.05f))
           .body("data.documents[0].$vectorize", is(notNullValue()));
-    }
-  }
-
-  @Nested
-  @Order(8)
-  @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-  @EnabledIfEnvironmentVariable(
-      named = AwsBedrockVectorSearchTest.AWS_ACCESS_KEY_ID,
-      matches = ".+")
-  class AwsBedrockVectorSearchTest {
-    static final String AWS_ACCESS_KEY_ID = "AWS_ACCESS_KEY_ID";
-    static final String AWS_SECRET_ID = "AWS_SECRET_ID";
-    final Map<String, ?> headers =
-        Map.of(
-            HttpConstants.AUTHENTICATION_TOKEN_HEADER_NAME,
-            getHeaders().get(HttpConstants.AUTHENTICATION_TOKEN_HEADER_NAME),
-            HttpConstants.EMBEDDING_AUTHENTICATION_ACCESS_ID_HEADER_NAME,
-            System.getenv(AWS_ACCESS_KEY_ID),
-            HttpConstants.EMBEDDING_AUTHENTICATION_SECRET_ID_HEADER_NAME,
-            System.getenv(AWS_SECRET_ID));
-
-    @Test
-    @Order(1)
-    public void happyPathVectorSearch() {
-
-      String json =
-          """
-          {
-              "createCollection": {
-                  "name": "aws_bedrock_vectorize",
-                  "options": {
-                      "vector": {
-                          "metric": "cosine",
-                          "service": {
-                              "provider": "bedrock",
-                              "modelName": "amazon.titan-embed-text-v2:0",
-                              "parameters": {
-                                  "region": "us-east-1"
-                              }
-                          }
-                      }
-                  }
-              }
-          }
-          """;
-      given()
-          .headers(headers)
-          .contentType(ContentType.JSON)
-          .body(json)
-          .when()
-          .post(NamespaceResource.BASE_PATH, namespaceName)
-          .then()
-          .statusCode(200)
-          .body("status.ok", is(1));
-    }
-
-    @Test
-    @Order(2)
-    public void insertVectorSearch() {
-      String json =
-          """
-          {
-             "insertOne": {
-                "document": {
-                    "_id": "1",
-                    "name": "Coded Cleats",
-                    "description": "ChatGPT integrated sneakers that talk to you",
-                    "$vectorize": "ChatGPT integrated sneakers that talk to you"
-                }
-             }
-          }
-          """;
-
-      given()
-          .headers(headers)
-          .contentType(ContentType.JSON)
-          .body(json)
-          .when()
-          .post(CollectionResource.BASE_PATH, namespaceName, "aws_bedrock_vectorize")
-          .then()
-          .statusCode(200)
-          .body("status.insertedIds[0]", is("1"))
-          .body("data", is(nullValue()))
-          .body("errors", is(nullValue()));
-
-      json =
-          """
-          {
-            "find": {
-              "filter" : {"_id" : "1"},
-              "projection": { "$vector": 1 }
-            }
-          }
-          """;
-
-      given()
-          .headers(headers)
-          .contentType(ContentType.JSON)
-          .body(json)
-          .when()
-          .post(CollectionResource.BASE_PATH, namespaceName, "aws_bedrock_vectorize")
-          .then()
-          .statusCode(200)
-          .body("errors", is(nullValue()))
-          .body("data.documents[0]._id", is("1"))
-          .body("data.documents[0].$vector", is(notNullValue()))
-          .body("data.documents[0].$vector", hasSize(1024));
-    }
-
-    @Test
-    @Order(4)
-    public void findOne() {
-      String json =
-          """
-                  {
-                    "findOne": {
-                      "sort" : {"$vectorize" : "ChatGPT integrated sneakers that talk to you"},
-                      "projection" : {"_id" : 1, "$vector" : 1, "$vectorize" : 1},
-                      "options" : {"includeSimilarity" : true}
-                    }
-                  }
-                  """;
-
-      given()
-          .headers(headers)
-          .contentType(ContentType.JSON)
-          .body(json)
-          .when()
-          .post(CollectionResource.BASE_PATH, namespaceName, "aws_bedrock_vectorize")
-          .then()
-          .statusCode(200)
-          .body("data.document._id", is("1"))
-          .body("data.document.$vector", is(notNullValue()))
-          .body("data.document.$vectorize", is(notNullValue()))
-          .body("data.document.$similarity", greaterThan(0.0f))
-          .body("errors", is(nullValue()));
     }
   }
 }
